Fix up asmcmd in 18c in addition to 12cR2
In #112 we added a fix-up for asmcmd errors that blocked grid installed in 12cR2, based on MOS note 2748316.1.  Trying an 18c install, we get the same error:

```
11-Mar-25 16:45 ASMCMD Background (PID = 1318138):  Successfully opened the pipe /var/tmp/.oracle/asmcmd_fg_1318129
NOTE: Executing /u01/app/18.0.0/grid/bin/kfod op=getclstype..
11-Mar-25 16:45 cluster mode - ASM cluster : Flex mode enabled
11-Mar-25 16:45 (Process=1318138) BG unhandled exception before exiting kgfnGetConnDetails requires 4 parameters at /u01/app/18.0.0/grid/lib/asmcmdbase.pm line 6096.
```

A close reading of note 2748316.1 shows that, while the fix was for 12.2, the underlying issue was only resolved in 19.1.  So here we add an 18c version of the fix-up, using 18c versions of the object files.