Add support for --ora-release in install-oracle.sh
`--ora-release` is used internally in the toolkit (and in
apply-patch.sh), but isn't formally taken as a parameter, so there's no
way to a user to specify a specific release to update to as part of
`install-oracle.sh`;  they'd need to install a base release and patch
up.
Secondarily, the terraform code accepts an `oracle-release` parameter,
which doesn't actually exist.

Here, we add the enw parameter, documentation, and terraform support.