Change os_disk to boot_disk in terraform
Change terminology to be more consistent with GCP's;  an OS disk could
be any disk managed by the OS, but a boot disk is more specifically the
disk that the OS boots from, and it's indeed a boot disk that we specify
here.

This is a breaking change;  taking advnatage of the Terraform refactor,
also a breaking change, to add it.