Make RMAN backups optional
History, the toolkit has required some sort of backup job to be
configured.  This change removes that requirement, allowing users to
specify a blank backup destination.

We still configure archivelog mode, to support future Data Guard
configurations, and to avoid the need for downtime if backups are
configured in the future.  If backups are not configued, a new "backup"
script `rman_delete_arch.sh` will delete accumulated archivelogs.