Update hugepages calculation
Hugepage size, currently, is derived from the default SGA calculation,
in turn based on memory size.  But it doesn't work is SGA size is
customized, be it in `all.yml` directly, or via the new command-line
version.

Updating the logic to use the `ora_sga_target` variable, which also
makes the calculation a bit simpler.