Enable serial port output logging for the control node
The transient control node does most of its work from the startup
script, which is logged via the serial port.  But after the VM is
deleted, the serial port output is no longer available.

The solution is to send the output to Cloud Logging, where it persists.

https://cloud.google.com/compute/docs/troubleshooting/viewing-serial-port-output#enable-stackdriver

Sample Terraform output:
https://gist.github.com/mfielding/28be7a0c31b8fc676c00b0f331faf7dc