Add `network` terraform variable
Adding a new variable `network`, that's required to support VPCs with
custom subnet creation mode.  In the default `auto` mode, network and
subnetwork are the same, so taking a single parameter is fine.  But with
`custom` mode, they aren't, so the caller needs to supply both.

`subnetwork` now defaults to the value of `network`.

Testcases:
https://gist.github.com/mfielding/1a08edf5295257e978dffa181fc120b7