Validate hugepage allocation
While testing memory allocations for #266, I realized that, if we
attempt to allocate more hugepages than the system can provide, the
command actually succeeds, allocating as many hugepages as it can.  This
is not a good outcome for us:  not only will the DB creation eventually
fail, but the install can hang due to swapping, since all free memory
has been allocated to hugepages.

This change runs the `sysctl` OS command to get the current value, and
asserts that it matches what we asked for.

Sample output:
https://gist.github.com/mfielding/8fc8c85a1bd62240e33c72341ee16748