Refactor commit trap and log setup
Modify the common lib to also set up the cleanup trap as part of it's
init, as one less thing the callers might forget.

And talking about forgetting, we had a setup_logging function, and never
called it.  Refactoring so that watch_logs does both the setup and
watching.  It's long, but there's really no need for callers to run part
of it.