Make disks larger for free edition test
Oracle 23c's DBCA requires at least 13GB of space on the RECO disk.
Also extending the data disk because, until #335 lands, the DATA and
RECO disks are classified as ASM disks and ignored, so the full install
actually goes on the boot disk.

Separately, make the free edition script executable in case someone
wants to run it directly using ./