Remove eval from Bash scripts where it was used to execute commands
## Change Description

Several scripts used `eval` unnecessarily, which can lead to security vulnerabilities.

This change removes the use of `eval` in these scripts, replacing it with an array containing the command and its arguments.

The array is then executed directly.

## Solution Overview

In general, replace this:

```bash
for PLAYBOOK in ${PB_LIST}; do
  ANSIBLE_COMMAND="${ANSIBLE_PLAYBOOK} ${ANSIBLE_PARAMS} ${ANSIBLE_EXTRA_PARAMS} ${PLAYBOOK}"
  echo
  echo "Running Ansible playbook: ${ANSIBLE_COMMAND}"
  eval "${ANSIBLE_COMMAND}"
done
```

with this:

```bash
for PLAYBOOK in ${PB_LIST}; do
  declare -a CMD_ARRAY=()
  CMD_ARRAY+="${ANSIBLE_PLAYBOOK}"
  CMD_ARRAY+=("${ANSIBLE_PARAMS}")
  CMD_ARRAY+=("${ANSIBLE_EXTRA_PARAMS}")
  CMD_ARRAY+=("${PLAYBOOK}")
  echo "echo Running Ansible playbook: ${CMD_ARRAY[@]}"
  ${CMD_ARRAY[@]}
done
```
That is a simplified version of what is now found in `install-oracle.sh`.

Other affected scripts have similar changes.

## Test Commands

In some cases a copy of the script was made, and the outputs from both methods of building the command line were compared.

In others, the script was run and the output examined to ensure it was correct.

With this change, the output of the ansible scripts is not really important, just that they run correctly.

There are detailed test instructions, logs and scripts in this document:

[oratk52-8-bash - detailed docs - oratk52-8-bash-cleanup.md](https://gist.github.com/jkstill/196edbd4665362f7069747e0ecd6a272)

## Expected Results

The expected result is that the ansible scripts run correctly, and that the output is as expected.

Tests of all scripts passed