Improve password checks
Update password checks to be more in line with the actual requirments
for SYS passwords as enforced by DBCA.  This is a higher standard than
the DEFAULT password policy.  This change will increase the complexity
standard for the Oracle metrics user, which formerly weren't validated
for complexity, only invalid characters.

Cleanups:

* Update terminology in task names and erorr messages to be consitent
  with the parameter names: database password and Oracle metrics
  password (rather than SYS and agent passwords).
* Remove redundant password check at user creation time
* Expand the allowed special characters, though we still exclude `&`
  because of its special meaning in sqlplus.
* Update task names to include the role as per the regular standard

Testcases:
https://gist.github.com/mfielding/5e0c99e83bb7dfd6a94b457f45a7f5bb