Add expiration time for OS login keys
Although we have a cleanup routing to attempt to clear out old,
unnecessary OS login SSH keys, it requires the cleanup handler to be
able to run, which doesn't always happen, especially when a test is
pre-empted.

This change explicitly sets a 24-hour key expiration to automatically
clean out keys that the cleanup might not catch.