Add start/stop script to cleanup
With the addition of #355 we now have a start/stop script for filesystem
installs. Here we add logic to attempt to use it to shut down the
database cleanup before a cleanup, and also to remove the service before
systemd attempts to restart it.

We attempt to preserve the existing semantics of the cleanup script
here:  to favor ignoring errors over complex "when" clauses, so as to
reduce the dependencies on getting parameters exactly right.

The systemd reload should also help with any other systemd services we
cleanup, like Oracle TFA.

Sample output from a test run (actually a re-run after a few bugfixes):
https://gist.github.com/mfielding/2b5e34df98d0bb7260c2055efb27b477